Extract separate GTFS-RT handler classes
# Conflicts:
#	application/src/main/java/org/opentripplanner/updater/trip/gtfs/GtfsRealTimeTripUpdateAdapter.java
