+++ application/src/main/java/org/opentripplanner/updater/trip/gtfs/ScheduledTripHandler.java
@@ -1,0 +1,117 @@
+package org.opentripplanner.updater.trip.gtfs;
+
+import static org.opentripplanner.updater.spi.UpdateErrorType.NO_SERVICE_ON_DATE;
+import static org.opentripplanner.updater.spi.UpdateErrorType.NO_UPDATES;
+import static org.opentripplanner.updater.spi.UpdateErrorType.TRIP_NOT_FOUND;
+
+import java.util.HashMap;
+import java.util.Map;
+import org.opentripplanner.core.model.id.FeedScopedId;
+import org.opentripplanner.transit.model.network.StopPattern;
+import org.opentripplanner.transit.model.network.TripPattern;
+import org.opentripplanner.transit.model.site.StopLocation;
+import org.opentripplanner.transit.model.timetable.RealTimeTripUpdate;
+import org.opentripplanner.transit.model.timetable.Trip;
+import org.opentripplanner.transit.service.TransitEditorService;
+import org.opentripplanner.updater.spi.UpdateException;
+import org.opentripplanner.updater.spi.UpdateSuccess;
+import org.opentripplanner.updater.trip.TimetableSnapshotManager;
+import org.opentripplanner.updater.trip.gtfs.model.TripUpdate;
+import org.opentripplanner.updater.trip.patterncache.TripPatternCache;
+
+class ScheduledTripHandler {
+
+  private final TransitEditorService transitEditorService;
+  private final TimetableSnapshotManager snapshotManager;
+  private final TripTimesUpdater tripTimesUpdater;
+  private final TripPatternCache tripPatternCache;
+
+  ScheduledTripHandler(
+    TransitEditorService transitEditorService,
+    TimetableSnapshotManager snapshotManager,
+    TripTimesUpdater tripTimesUpdater,
+    TripPatternCache tripPatternCache
+  ) {
+    this.transitEditorService = transitEditorService;
+    this.snapshotManager = snapshotManager;
+    this.tripTimesUpdater = tripTimesUpdater;
+    this.tripPatternCache = tripPatternCache;
+  }
+
+  UpdateSuccess handle(
+    TripUpdate tripUpdate,
+    ForwardsDelayPropagationType forwardsDelayPropagationType,
+    BackwardsDelayPropagationType backwardsDelayPropagationType
+  ) throws UpdateException {
+    final TripPattern pattern = getPatternForTripId(tripUpdate.tripId());
+
+    if (pattern == null) {
+      throw UpdateException.of(tripUpdate.tripId(), TRIP_NOT_FOUND);
+    }
+
+    if (tripUpdate.stopTimeUpdates().isEmpty()) {
+      throw UpdateException.of(tripUpdate.tripId(), NO_UPDATES);
+    }
+
+    var serviceId = transitEditorService.getTrip(tripUpdate.tripId()).getServiceId();
+    var serviceDates = transitEditorService
+      .getCalendarService()
+      .getServiceDatesForServiceId(serviceId);
+    if (!serviceDates.contains(tripUpdate.serviceDate())) {
+      throw UpdateException.of(tripUpdate.tripId(), NO_SERVICE_ON_DATE);
+    }
+
+    var tripTimesPatch = tripTimesUpdater.createUpdatedTripTimesFromGtfsRt(
+      pattern.getScheduledTimetable(),
+      tripUpdate,
+      forwardsDelayPropagationType,
+      backwardsDelayPropagationType
+    );
+
+    var updatedPickup = tripTimesPatch.updatedPickup();
+    var updatedDropoff = tripTimesPatch.updatedDropoff();
+    var replacedStopIndices = tripTimesPatch.replacedStopIndices();
+    var updatedTripTimes = tripTimesPatch.tripTimes();
+
+    Map<Integer, StopLocation> newStops = new HashMap<>();
+    for (var entry : replacedStopIndices.entrySet()) {
+      var stop = transitEditorService.getRegularStop(
+        new FeedScopedId(tripUpdate.tripId().getFeedId(), entry.getValue())
+      );
+      if (stop != null) {
+        newStops.put(entry.getKey(), stop);
+      }
+    }
+
+    // If there are stops with different pickup / drop off, or replaced stops, we need to change the pattern from the scheduled one
+    if (!updatedPickup.isEmpty() || !updatedDropoff.isEmpty() || !newStops.isEmpty()) {
+      StopPattern newStopPattern = pattern
+        .copyPlannedStopPattern()
+        .updatePickups(updatedPickup)
+        .updateDropoffs(updatedDropoff)
+        .replaceStops(newStops)
+        .build();
+
+      final Trip trip = transitEditorService.getTrip(tripUpdate.tripId());
+      final TripPattern newPattern = tripPatternCache.getOrCreateTripPattern(newStopPattern, trip);
+
+      return snapshotManager.updateBuffer(
+        RealTimeTripUpdate.of(newPattern, updatedTripTimes, tripUpdate.serviceDate())
+          .withRevertPreviousRealTimeUpdates(true)
+          .withHideTripInScheduledPattern(pattern)
+          .build()
+      );
+    } else {
+      return snapshotManager.updateBuffer(
+        RealTimeTripUpdate.of(pattern, updatedTripTimes, tripUpdate.serviceDate())
+          .withRevertPreviousRealTimeUpdates(true)
+          .build()
+      );
+    }
+  }
+
+  private TripPattern getPatternForTripId(FeedScopedId tripId) {
+    Trip trip = transitEditorService.getTrip(tripId);
+    return transitEditorService.findPattern(trip);
+  }
+}
